style: ディレクトリ構成をわかりやすく変更
